Refactor: Use refinements for String extension
